Separate climb subsystem into elevator and winch subsystems
