+++ 2022Robot/src/main/java/frc/robot/commands/AutoRoutines/TraverseRungs.java
@@ -1,10 +1,10 @@
 package frc.robot.commands.AutoRoutines;
 
 import edu.wpi.first.wpilibj2.command.CommandBase;
-import frc.robot.subsystems.ClimbSubsystem;
+import frc.robot.subsystems.ElevatorSubsystem;
 
 public class TraverseRungs extends CommandBase {
-    public TraverseRungs(ClimbSubsystem climbSubsystem) {
+    public TraverseRungs(ElevatorSubsystem climbSubsystem) {
         // to be programmed later
     }
 }
