+++ 2022Robot/src/main/java/frc/robot/commands/ExtendElevator.java
@@ -2,29 +2,29 @@ package frc.robot.commands;
 
 import edu.wpi.first.wpilibj2.command.CommandBase;
 import frc.robot.Constants;
-import frc.robot.subsystems.ClimbSubsystem;
+import frc.robot.subsystems.ElevatorSubsystem;
 
 public class ExtendElevator extends CommandBase {
-    private final ClimbSubsystem m_climbSubsystem;
+    private final ElevatorSubsystem m_elevatorSubsystem;
 
-    public ExtendElevator(ClimbSubsystem climbSubsystem) {
-        m_climbSubsystem = climbSubsystem;
-        addRequirements(climbSubsystem);
+    public ExtendElevator(ElevatorSubsystem elevatorSubsystem) {
+        m_elevatorSubsystem = elevatorSubsystem;
+        addRequirements(elevatorSubsystem);
     }
 
     @Override
     public void execute() {
-        m_climbSubsystem.setElevator(Constants.ClimbConstants.kExtendElevatorSpeed);
+        m_elevatorSubsystem.setElevator(Constants.ElevatorConstants.kExtendElevatorSpeed);
     }
 
     @Override
     public void initialize() {
-        m_climbSubsystem.stopElevator();
+        m_elevatorSubsystem.stopElevator();
     }
 
     @Override
     public void end(boolean interrupted) {
-        m_climbSubsystem.stopElevator();
+        m_elevatorSubsystem.stopElevator();
     }
 
     @Override
